Added the beginnings of an analytical check
When completed, Check() will give a weighted score out of 100 (or
whatever) based on data. For example, Bz in the red will be worth 25
points, while yellow will be worth 10, and green, -10. Kp will be given
less emphasis, as personal experience shows a Kp 7 storm means squat
when an aurora is already dying down.